refactor helper functions
